Examples: consolidate Geant4/DD4hep code
