+++ Examples/Algorithms/Geant4/src/DD4hepToG4Svc.cpp
@@ -6,9 +6,9 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "ACTFW/Plugins/DD4hepG4/DD4hepToG4Svc.hpp"
+#include "ActsExamples/Geant4/DD4hepToG4Svc.hpp"
 
-#include "ACTFW/Plugins/DD4hepG4/GeoConstruction.hpp"
+#include "ActsExamples/Geant4/GeoConstruction.hpp"
 
 FW::DD4hepG4::DD4hepToG4Svc::DD4hepToG4Svc(
     const FW::DD4hepG4::DD4hepToG4Svc::Config& cfg)
