+++ Examples/Algorithms/Geant4/src/GeoConstruction.cpp
@@ -6,7 +6,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "ACTFW/Plugins/DD4hepG4/GeoConstruction.hpp"
+#include "ActsExamples/Geant4/GeoConstruction.hpp"
 
 #include <stdexcept>
