+++ Examples/Algorithms/Geant4/src/MMSteppingAction.cpp
@@ -6,13 +6,13 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "ACTFW/Plugins/Geant4/MMSteppingAction.hpp"
+#include "MMSteppingAction.hpp"
 
+#include <G4Material.hh>
+#include <G4Step.hh>
 #include <stdexcept>
 
 #include "Acts/Utilities/Units.hpp"
-#include "G4Material.hh"
-#include "G4Step.hh"
 
 FW::Geant4::MMSteppingAction* FW::Geant4::MMSteppingAction::fgInstance =
     nullptr;
