+++ Examples/Algorithms/Geant4/src/MMSteppingAction.hpp
@@ -12,12 +12,12 @@
 
 #pragma once
 
+#include <G4UserSteppingAction.hh>
+#include <globals.hh>
 #include <vector>
 
 #include "ACTFW/EventData/SimHit.hpp"
 #include "Acts/Propagator/MaterialInteractor.hpp"
-#include "G4UserSteppingAction.hh"
-#include "globals.hh"
 
 namespace FW {
 namespace Geant4 {
