+++ Examples/Algorithms/MaterialMapping/CMakeLists.txt
@@ -8,21 +8,6 @@ target_link_libraries(
   ActsExamplesMaterialMapping
   PUBLIC ActsCore ActsExamplesFramework)
 
-if(ACTS_BUILD_EXAMPLES_DD4HEP AND ACTS_BUILD_EXAMPLES_GEANT4)
-  target_sources(
-    ActsExamplesMaterialMapping
-    PRIVATE src/GeantinoRecording.cpp)
-  target_compile_definitions(
-    ActsExamplesMaterialMapping
-    PUBLIC ${Geant4_DEFINITIONS})
-  target_include_directories(
-    ActsExamplesMaterialMapping
-    SYSTEM PUBLIC ${Geant4_INCLUDE_DIRS})
-  target_link_libraries(
-    ActsExamplesMaterialMapping
-    PUBLIC ActsExamplesGeant4 ActsExamplesDD4hepG4 ${Geant4_LIBRARIES})
-endif()
-
 install(
   TARGETS ActsExamplesMaterialMapping
   LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR})
