+++ Examples/Algorithms/MaterialMapping/include/ACTFW/MaterialMapping/MaterialMappingOptions.hpp
@@ -8,10 +8,10 @@
 
 #pragma once
 
-#include <Acts/Utilities/Units.hpp>
 #include <iostream>
 
 #include "ACTFW/Utilities/Options.hpp"
+#include "Acts/Utilities/Units.hpp"
 
 namespace po = boost::program_options;
